Remove InputLabel from StandardInputWithLabel.tsx, add change label with placeholder prop then make Input component to take placeholder. Update Login.tsx to match the changes
